Separate providers from custom hooks and context creation
