+++ src/components/common/Navigation.jsx
@@ -1,6 +1,6 @@
 import { NavLink, useNavigate } from "react-router-dom";
 import { Container, Nav, Navbar, Button } from "react-bootstrap";
-import { useAuth } from "../../contexts/AuthContext";
+import { useAuth } from "../../contexts/auth";
 
 import reactLogo from "/react-logo.png";
