+++ src/contexts/auth/index.jsx
@@ -1,0 +1,3 @@
+export { AuthContext } from "./AuthContext";
+export { AuthProvider } from "./AuthProvider";
+export { useAuth } from "./useAuth";
