+++ src/contexts/functions/FunctionsContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const FunctionsContext = createContext(null);
