+++ src/contexts/functions/index.jsx
@@ -1,0 +1,3 @@
+export { FunctionsContext } from "./FunctionsContext";
+export { FunctionsProvider } from "./FunctionsProvider";
+export { useFunctions } from "./useFunctions";
