+++ src/contexts/movies/MoviesContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const MoviesContext = createContext(null);
