+++ src/contexts/movies/index.jsx
@@ -1,0 +1,3 @@
+export { MoviesContext } from "./MoviesContext";
+export { MoviesProvider } from "./MoviesProvider";
+export { useMovies } from "./useMovies";
