+++ src/contexts/movies/useMovies.jsx
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { MoviesContext } from "./MoviesContext";
+
+export const useMovies = () => {
+  const context = useContext(MoviesContext);
+  if (!context) {
+    throw new Error("useMovies must be used within MoviesProvider");
+  }
+  return context;
+};
