+++ src/main.jsx
@@ -1,8 +1,8 @@
 import { createRoot } from "react-dom/client";
 import "bootstrap/dist/css/bootstrap.min.css";
-import { MoviesProvider } from "./contexts/MoviesContext";
-import { FunctionsProvider } from "./contexts/FunctionsContext";
-import { AuthProvider } from "./contexts/AuthContext";
+import { MoviesProvider } from "./contexts/movies";
+import { FunctionsProvider } from "./contexts/functions";
+import { AuthProvider } from "./contexts/auth";
 import App from "./App.jsx";
 import "./styles/custom.css";
