+++ src/pages/Login.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Form, Container, Card } from "react-bootstrap";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../contexts/auth";
 
 import LoginFields from "../components/login/LoginFields";
 import LoginActions from "../components/login/LoginActions";
